Expand stencil state to 64 bits and update handling

+++ src/vrhi_internal.h
@@ -277,7 +277,7 @@ void vhWriteStateToGlobalUniform( const vhState& state, vhGlobalUniform& out );
 void vhWriteStateToWorldUniform( const vhState& state, vhWorldUniform& out );
 nvrhi::PrimitiveType vhTranslatePrimitiveType( uint64_t stateFlags );
 nvrhi::BlendState vhTranslateBlendState( uint64_t stateFlags );
-nvrhi::DepthStencilState vhTranslateDepthStencilState( uint64_t stateFlags, uint32_t frontStencil, uint32_t backStencil );
+nvrhi::DepthStencilState vhTranslateDepthStencilState( uint64_t stateFlags, uint64_t frontStencil, uint64_t backStencil );
 nvrhi::RasterState vhTranslateRasterState( uint64_t stateFlags );
 nvrhi::VariableShadingRate vhTranslateShadingRate( uint32_t rate );
 nvrhi::ShadingRateCombiner vhTranslateShadingRateCombiner( uint32_t combiner );
